fix: use 32-bit Codex RPC metrics fields

+++ themes/codex-status/include/zmk/codex_metrics.h
@@ -4,8 +4,8 @@
 
 struct zmk_codex_metrics {
     uint8_t five_hour_used_percent;
-    uint64_t today_total_tokens;
-    uint64_t updated_at;
+    uint32_t today_total_tokens;
+    uint32_t updated_at;
 };
 
 struct zmk_codex_metrics zmk_codex_metrics_get(void);
